+++ packages/driver/cypress/e2e/commands/sessions/sessions.cy.js
@@ -8,10 +8,19 @@ before(() => {
 const expectCurrentSessionData = async (obj) => {
   return Cypress.session.getCurrentSessionData()
   .then((result) => {
-    cy.log(result)
     expect(result.cookies.map((v) => v.name)).members(obj.cookies || [])
     expect(result.localStorage).deep.members(obj.localStorage || [])
     expect(result.sessionStorage).deep.members(obj.sessionStorage || [])
+  })
+}
+
+const clearAllSavedSessions = () => {
+  // clear all sessions only sets hydrated: false and re-using a session id
+  // with new setup / validation fn isn't updated/applied
+  Cypress.state('activeSessions', {})
+  cy.log('Cypress.session.clearAllSavedSessions()')
+  cy.then(async () => {
+    return Cypress.session.clearAllSavedSessions()
   })
 }
 
@@ -137,13 +146,7 @@ describe('cy.session', { retries: 0 }, () => {
 
     const setupTestContext = () => {
       resetMocks()
-
-      // clear all sessions only sets hydrated: false and re-using a session id
-      // with new setup / validation fn isn't updated/applied
-      Cypress.state('activeSessions', {})
-      cy.log('Cypress.session.clearAllSavedSessions()')
-      Cypress.session.clearAllSavedSessions()
-
+      clearAllSavedSessions()
       cy.on('log:added', (attrs, log) => {
         if (attrs.name === 'session' || attrs.name === 'sessions_manager' || attrs.name === 'page load' || attrs.alias?.includes('setupSession') || attrs.alias?.includes('validateSession')) {
           logs.push(log)
@@ -791,6 +794,7 @@ describe('cy.session', { retries: 0 }, () => {
     }
 
     beforeEach(() => {
+      clearAllSavedSessions()
       cy.on('log:added', handleAddLog)
       cy.on('fail', (err) => {
         cy.off('log:added', handleAddLog)
@@ -800,7 +804,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws error when experimentalSessionAndOrigin not enabled', { experimentalSessionAndOrigin: false, experimentalSessionSupport: false }, (done) => {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastSessionLog.get('error')).to.eq(err)
         expect(lastSessionLog.get('state')).to.eq('failed')
@@ -814,7 +818,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws error when experimentalSessionSupport is enabled through test config', { experimentalSessionAndOrigin: false, experimentalSessionSupport: true }, (done) => {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastSessionLog.get('error')).to.eq(err)
         expect(lastSessionLog.get('state')).to.eq('failed')
@@ -830,7 +834,7 @@ describe('cy.session', { retries: 0 }, () => {
     it('throws error when experimentalSessionSupport is enabled through Cypress.config', { experimentalSessionAndOrigin: false }, (done) => {
       Cypress.config('experimentalSessionSupport', true)
 
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         Cypress.config('experimentalSessionSupport', false)
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
@@ -844,7 +848,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws when sessionId argument was not provided', function (done) {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
@@ -858,7 +862,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws when sessionId argument is not an object', function (done) {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
@@ -872,7 +876,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws when options argument is provided and is not an object', function (done) {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
@@ -886,7 +890,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws when options argument has an invalid option', function (done) {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
@@ -900,7 +904,7 @@ describe('cy.session', { retries: 0 }, () => {
     })
 
     it('throws when options argument has an option with an invalid type', function (done) {
-      cy.on('fail', (err) => {
+      cy.once('fail', (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
@@ -913,17 +917,57 @@ describe('cy.session', { retries: 0 }, () => {
       cy.session('some-session', () => {}, { validate: 2 })
     })
 
-    it('throws when multiple session calls with same sessionId but different options', function (done) {
+    it('throws when setup function is not provided and existing session is not found', function (done) {
+      cy.once('fail', (err) => {
+        expect(lastLog.get('error')).to.eq(err)
+        expect(lastLog.get('state')).to.eq('failed')
+        expect(err.message).to.eq('No session is defined with the name\n  **some-session**\nIn order to use `cy.session()`, provide a `setup` as the second argument:\n\n`cy.session(id, setup)`')
+        expect(err.docsUrl).to.eq('https://on.cypress.io/session')
+
+        done()
+      })
+
+      cy.session('some-session')
+    })
+
+    it('throws when setup function is not provided and global session is registered', function (done) {
+      cy.once('fail', (err) => {
+        expect(lastSessionLog).to.eq(lastLog)
+        expect(lastLog.get('error')).to.eq(err)
+        expect(lastLog.get('state')).to.eq('failed')
+        expect(err.message).to.eq('In order to restore a global `cy.session()`, provide a `setup` as the second argument:\n\n`cy.session(id, setup, { cacheAcrossSpecs: true })`')
+        expect(err.docsUrl).to.eq('https://on.cypress.io/session')
+
+        done()
+      })
+
+      cy.session('some-session-2', () => {}, { cacheAcrossSpecs: true })
+      .then(() => {
+        Cypress.state('activeSessions', {})
+      }).then(async () => {
+        await Cypress.backend('get:session', 'some-session-2').then((sessionDetails) => {
+          Cypress.state('activeSessions', { 'some-session-2': sessionDetails })
+        })
+      })
+
+      cy.session('some-session-2')
+    })
+
+    it('throws when multiple session calls with same sessionId but different setup', function (done) {
       cy.once('fail', async (err) => {
         expect(lastSessionLog).to.eq(lastLog)
         expect(lastLog.get('error')).to.eq(err)
         expect(lastLog.get('state')).to.eq('failed')
-        expect(err.message).to.eq('You may not call `cy.session()` with a previously used name and different options. If you want to specify different options, please use a unique name other than **duplicate-session**.')
+        expect(err.message).to.eq('This session already exists. You may not create a new session with a previously used identifier. If you want to create a new session with a different setup function, please call `cy.session()` with a unique identifier other than **duplicate-session**.')
         expect(err.docsUrl).to.eq('https://on.cypress.io/session')
 
-        await expectCurrentSessionData({
-          localStorage: [{ origin: baseUrl, value: { one: 'value' } }],
-        })
+        try {
+          await expectCurrentSessionData({
+            localStorage: [{ origin: baseUrl, value: { one: 'value' } }],
+          })
+        } catch (err) {
+          done(err)
+        }
 
         done()
       })
@@ -939,20 +983,39 @@ describe('cy.session', { retries: 0 }, () => {
       })
     })
 
-    describe('setup function failures', () => {
-      it('throws when setup function is not provided and existing session is not found', function (done) {
-        cy.on('fail', (err) => {
-          expect(lastLog.get('error')).to.eq(err)
-          expect(lastLog.get('state')).to.eq('failed')
-          expect(err.message).to.eq('No session is defined with the name\n  **some-session**\nIn order to use `cy.session()`, provide a `setup` as the second argument:\n\n`cy.session(id, setup)`')
-          expect(err.docsUrl).to.eq('https://on.cypress.io/session')
+    it('throws when multiple session calls with same sessionId but different validate opt', function (done) {
+      cy.once('fail', async (err) => {
+        expect(lastSessionLog).to.eq(lastLog)
+        expect(lastLog.get('error')).to.eq(err)
+        expect(lastLog.get('state')).to.eq('failed')
+        expect(err.message).to.eq('This session already exists. You may not create a new session with a previously used identifier. If you want to create a new session with a different validate function, please call `cy.session()` with a unique identifier other than **duplicate-sess**.')
+        expect(err.docsUrl).to.eq('https://on.cypress.io/session')
 
-          done()
-        })
-
-        cy.session('some-session')
+        done()
       })
 
+      cy.session('duplicate-sess', () => {}, { validate: () => {} })
+
+      cy.session('duplicate-sess', () => {}, { validate: () => { /* do something */ } })
+    })
+
+    it('throws when multiple session calls with same sessionId but different cacheAcrossSpec opt', function (done) {
+      cy.once('fail', async (err) => {
+        expect(lastSessionLog).to.eq(lastLog)
+        expect(lastLog.get('error')).to.eq(err)
+        expect(lastLog.get('state')).to.eq('failed')
+        expect(err.message).to.eq('This session already exists. You may not create a new session with a previously used identifier. If you want to create a new session with a different persistence, please call `cy.session()` with a unique identifier other than **duplicate-sess**.')
+        expect(err.docsUrl).to.eq('https://on.cypress.io/session')
+
+        done()
+      })
+
+      cy.session('duplicate-sess', () => {}, { validate: () => {} })
+
+      cy.session('duplicate-sess', () => {}, { validate: () => {}, cacheAcrossSpecs: true })
+    })
+
+    describe('setup function failures', () => {
       it('throws when setup function has a failing Cypress command', function (done) {
         cy.once('fail', (err) => {
           expect(lastLog.get('error')).to.eq(err)
@@ -983,11 +1046,11 @@ describe('cy.session', { retries: 0 }, () => {
       })
     })
 
-    describe.skip('options.validate failures', () => {
+    describe('options.validate failures', () => {
       const errorHookMessage = 'This error occurred in a session validate hook after initializing the session. Because validation failed immediately after session setup we failed the test.'
 
       it('throws when options.validate has a failing Cypress command', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).contain('Expected to find element: `#does_not_exist`')
           expect(err.message).contain(errorHookMessage)
           expect(err.codeFrame).exist
@@ -1005,7 +1068,7 @@ describe('cy.session', { retries: 0 }, () => {
       })
 
       it('throws when options.validate throws an error', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).contain('validate error')
           expect(err.message).contain(errorHookMessage)
           expect(err.codeFrame).exist
@@ -1022,7 +1085,7 @@ describe('cy.session', { retries: 0 }, () => {
       })
 
       it('throws when options.validate rejects', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).contain('validate error')
           expect(err.message).contain(errorHookMessage)
           expect(err.codeFrame).exist
@@ -1040,7 +1103,7 @@ describe('cy.session', { retries: 0 }, () => {
       })
 
       it('throws when options.validate returns false', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).to.contain('Your `cy.session` **validate** callback returned false.')
           expect(err.message).contain(errorHookMessage)
           expect(err.codeFrame).exist
@@ -1058,7 +1121,7 @@ describe('cy.session', { retries: 0 }, () => {
       })
 
       it('throws when options.validate resolves false', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).to.contain('Your `cy.session` **validate** callback resolved false.')
           expect(err.message).contain(errorHookMessage)
           expect(err.codeFrame).exist
@@ -1080,7 +1143,7 @@ describe('cy.session', { retries: 0 }, () => {
       // make error collapsible by default
 
       it('throws when options.validate returns Chainer<false>', (done) => {
-        cy.on('fail', (err) => {
+        cy.once('fail', (err) => {
           expect(err.message).to.contain('Your `cy.session` **validate** callback resolved false.')
           expect(err.message).contain(errorHookMessage)
           done()