+++ packages/driver/cypress/e2e/e2e/origin/commands/cookies.cy.ts
@@ -8,7 +8,7 @@ describe('cy.origin cookies', () => {
     })
 
     it('.getCookie(), .getCookies(), and .setCookie()', () => {
-      cy.origin('http://foobar.com:3500', () => {
+      cy.origin('http://www.foobar.com:3500', () => {
         cy.getCookies().should('be.empty')
 
         cy.setCookie('foo', 'bar')
@@ -19,7 +19,7 @@ describe('cy.origin cookies', () => {
     })
 
     it('.clearCookie()', () => {
-      cy.origin('http://foobar.com:3500', () => {
+      cy.origin('http://www.foobar.com:3500', () => {
         cy.setCookie('foo', 'bar')
         cy.getCookie('foo').should('not.be.null')
         cy.clearCookie('foo')
@@ -28,7 +28,7 @@ describe('cy.origin cookies', () => {
     })
 
     it('.clearCookies()', () => {
-      cy.origin('http://foobar.com:3500', () => {
+      cy.origin('http://www.foobar.com:3500', () => {
         cy.setCookie('foo', 'bar')
         cy.setCookie('faz', 'baz')
 
@@ -58,7 +58,7 @@ describe('cy.origin cookies', () => {
           expect(logs[0].get('state')).to.eq('failed')
           expect(logs[0].get('name')).to.eq('getCookie')
           expect(logs[0].get('message')).to.eq('foo')
-          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
+          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://www.foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
 
           done()
         })
@@ -73,7 +73,7 @@ describe('cy.origin cookies', () => {
           expect(logs[0].get('state')).to.eq('failed')
           expect(logs[0].get('name')).to.eq('getCookies')
           expect(logs[0].get('message')).to.eq('')
-          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
+          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://www.foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
 
           done()
         })
@@ -88,7 +88,7 @@ describe('cy.origin cookies', () => {
           expect(logs[0].get('state')).to.eq('failed')
           expect(logs[0].get('name')).to.eq('setCookie')
           expect(logs[0].get('message')).to.eq('foo, bar')
-          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
+          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://www.foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
 
           done()
         })
@@ -103,7 +103,7 @@ describe('cy.origin cookies', () => {
           expect(logs[0].get('state')).to.eq('failed')
           expect(logs[0].get('name')).to.eq('clearCookie')
           expect(logs[0].get('message')).to.eq('foo')
-          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
+          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://www.foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
 
           done()
         })
@@ -118,7 +118,7 @@ describe('cy.origin cookies', () => {
           expect(logs[0].get('state')).to.eq('failed')
           expect(logs[0].get('name')).to.eq('clearCookies')
           expect(logs[0].get('message')).to.eq('')
-          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
+          expect(err.message).to.eq('Timed out retrying after 100ms: The command was expected to run against origin `http://localhost:3500` but the application is at origin `http://www.foobar.com:3500`.\n\nThis commonly happens when you have either not navigated to the expected origin or have navigated away unexpectedly.')
 
           done()
         })
@@ -140,7 +140,7 @@ describe('cy.origin cookies', () => {
       })
 
       it('.getCookie()', () => {
-        cy.origin('http://foobar.com:3500', () => {
+        cy.origin('http://www.foobar.com:3500', () => {
           cy.getCookies().should('be.empty')
           cy.setCookie('foo', 'bar')
           cy.getCookie('foo').its('value').should('equal', 'bar')
@@ -161,7 +161,7 @@ describe('cy.origin cookies', () => {
       })
 
       it('.getCookies()', () => {
-        cy.origin('http://foobar.com:3500', () => {
+        cy.origin('http://www.foobar.com:3500', () => {
           cy.getCookies().should('be.empty')
 
           cy.setCookie('foo', 'bar')
@@ -189,7 +189,7 @@ describe('cy.origin cookies', () => {
       })
 
       it('.setCookie()', () => {
-        cy.origin('http://foobar.com:3500', () => {
+        cy.origin('http://www.foobar.com:3500', () => {
           cy.getCookies().should('be.empty')
 
           cy.setCookie('foo', 'bar')
@@ -210,7 +210,7 @@ describe('cy.origin cookies', () => {
       })
 
       it('.clearCookie()', () => {
-        cy.origin('http://foobar.com:3500', () => {
+        cy.origin('http://www.foobar.com:3500', () => {
           cy.setCookie('foo', 'bar')
           cy.getCookie('foo').should('not.be.null')
           cy.clearCookie('foo')
@@ -232,7 +232,7 @@ describe('cy.origin cookies', () => {
       })
 
       it('.clearCookies()', () => {
-        cy.origin('http://foobar.com:3500', () => {
+        cy.origin('http://www.foobar.com:3500', () => {
           cy.setCookie('foo', 'bar')
           cy.setCookie('faz', 'baz')