+++ packages/driver/cypress/e2e/e2e/origin/logging.cy.ts
@@ -11,7 +11,7 @@ describe('cy.origin logging', () => {
     cy.visit('/fixtures/primary-origin.html')
     cy.get('a[data-cy="cross-origin-secondary-link"]').click()
 
-    cy.origin('http://foobar.com:3500', () => {
+    cy.origin('http://www.foobar.com:3500', () => {
       cy.get('[data-cy="dom-check"]').invoke('text')
     })
 
@@ -70,7 +70,7 @@ describe('cy.origin logging', () => {
 
     const options = { args: { div: Cypress.$('div') } }
 
-    cy.origin('http://foobar.com:3500', options, () => {})
+    cy.origin('http://www.foobar.com:3500', options, () => {})
   })
 
   it('groups callback commands when failing with inner command failure', (done) => {
@@ -99,7 +99,7 @@ describe('cy.origin logging', () => {
     cy.visit('/fixtures/primary-origin.html')
     cy.get('a[data-cy="cross-origin-secondary-link"]').click()
 
-    cy.origin('http://foobar.com:3500', () => {
+    cy.origin('http://www.foobar.com:3500', () => {
       cy.get('[data-cy="dom-check"]').invoke('text')
       cy.get('#does-not-exist', { timeout: 1 })
     })
@@ -131,7 +131,7 @@ describe('cy.origin logging', () => {
     cy.visit('/fixtures/primary-origin.html')
     cy.get('a[data-cy="cross-origin-secondary-link"]').click()
 
-    cy.origin('http://foobar.com:3500', () => {
+    cy.origin('http://www.foobar.com:3500', () => {
       cy.get('[data-cy="dom-check"]').invoke('text').then(() => {
         setTimeout(() => {
           throw new Error('async error')