+++ packages/driver/src/cy/commands/sessions/manager.ts
@@ -124,6 +124,7 @@ export default class SessionsManager {
         setup: options.setup,
         hydrated: false,
         validate: options.validate,
+        cacheAcrossSpecs: !!options.cacheAcrossSpecs,
       }
 
       this.setActiveSession({ [sess_state.id]: sess_state })
@@ -133,6 +134,7 @@ export default class SessionsManager {
 
     clearAllSavedSessions: async () => {
       this.clearActiveSessions()
+      this.registeredSessions.clear()
       const clearAllSessions = true
 
       return this.Cypress.backend('clear:sessions', clearAllSessions)
@@ -158,7 +160,7 @@ export default class SessionsManager {
 
       // persist the session to the server. Only matters in openMode OR if there's a top navigation on a future test.
       // eslint-disable-next-line no-console
-      return this.Cypress.backend('save:session', { ...data, setup: data.setup.toString() }).catch(console.error)
+      return this.Cypress.backend('save:session', { ...data, setup: data.setup.toString(), validate: data.validate?.toString() }).catch(console.error)
     },
 
     setSessionData: async (data) => {