+++ system-tests/project-fixtures/angular/src/app/components/parent-providers.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  template: `
+    <app-child-providers></app-child-providers>
+    <app-another-child></app-another-child>`
+})
+export class ParentProvidersComponent {}